+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeClient.java
@@ -321,4 +321,14 @@ public class BraintreeClient {
     Context getApplicationContext() {
         return applicationContext;
     }
+
+    /**
+     * For clients using a {@link ClientTokenProvider}, call this method to invalidate the existing,
+     * cached client token. A new client token will be fetched by the SDK when it is needed.
+     *
+     * For clients not using a {@link ClientTokenProvider}, this method does nothing.
+     */
+    public void invalidateClientToken() {
+        authorizationLoader.invalidateClientToken();
+    }
 }