+++ TestUtils/src/main/java/com/braintreepayments/api/MockAuthorizationLoaderBuilder.java
@@ -4,8 +4,6 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 
-import android.content.Context;
-
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;